requirements.txt
# API Framework
fastapi==0.110.0
uvicorn==0.27.1

# Machine Learning & Clustering
scikit-learn==1.4.1.post1
numpy==1.26.4

# Embeddings & NLP
sentence-transformers==2.5.1

# Vector Database
faiss-cpu==1.8.0

# Exploration & Visualization (For your cluster justifications)
matplotlib==3.8.3
jupyter==1.0.0